requirements.txt
Flask==2.0.3
python-dotenv==1.0.1
stripe==10.9.0
aws-wsgi==0.2.7